Renamed StateController.js to Router.js
