+++ NavigationReact/sample/hotloader/index.js
@@ -1,13 +1,13 @@
 var Navigation = require('navigation');
-var StateController = require('./StateController');
+var Router = require('./Router');
 
 var stateNavigator = new Navigation.StateNavigator();
-StateController.configure(stateNavigator);
+Router.configure(stateNavigator);
 stateNavigator.start();
 
 if (module.hot) {
-    module.hot.accept('./StateController', function() {
-        require('./StateController').configure(stateNavigator);
+    module.hot.accept('./Router', function() {
+        require('./Router').configure(stateNavigator);
         stateNavigator.navigate(stateNavigator.stateContext.state.key, stateNavigator.stateContext.includeCurrentData({}));
     });
 }
